refactor app.py (2/7): extract db helpers to helpers/db.py
get_db, slugify, unique_slug, et_now move to helpers/db.py. The
DB_FILE constant is redefined locally in helpers/db.py from the
COCKPIT_REPO_ROOT env var (same fallback chain as app.py — production
unaffected, local dev still uses .env override).

Smoke-tested: /below-deck, /command-deck/, /command-deck/projects/
all 200 (these all open SQLite via get_db).
