+++ app.py
@@ -1052,48 +1052,9 @@ def today_complete():
 	conn.close()
 	return jsonify({'success': True})
 
-# ---- COMMAND DECK HELPERS ----
+# ---- COMMAND DECK HELPERS ---- (moved to helpers/db.py)
 
-def get_db():
-	"""Open a SQLite connection with WAL mode and foreign keys on."""
-	conn = sqlite3.connect(DB_FILE)
-	conn.row_factory = sqlite3.Row
-	conn.execute("PRAGMA foreign_keys = ON")
-	conn.execute("PRAGMA journal_mode = WAL")
-	return conn
-
-
-def slugify(text):
-	"""Turn a project title into a URL-safe slug."""
-	text = text.lower().strip()
-	text = re.sub(r'[^\w\s-]', '', text)
-	text = re.sub(r'[\s_-]+', '-', text)
-	text = re.sub(r'^-+|-+$', '', text)
-	return text or 'project'
-
-
-def unique_slug(title, conn, exclude_id=None):
-	"""Generate a unique slug, appending -2, -3 etc. if needed."""
-	base = slugify(title)
-	slug = base
-	n = 2
-	while True:
-		if exclude_id:
-			row = conn.execute(
-				'SELECT id FROM projects WHERE slug = ? AND id != ?', (slug, exclude_id)
-			).fetchone()
-		else:
-			row = conn.execute('SELECT id FROM projects WHERE slug = ?', (slug,)).fetchone()
-		if not row:
-			return slug
-		slug = f'{base}-{n}'
-		n += 1
-
-
-def et_now():
-	"""Current time as ISO string in US/Eastern."""
-	eastern = pytz.timezone('US/Eastern')
-	return datetime.now(eastern).isoformat()
+from helpers.db import get_db, slugify, unique_slug, et_now
 
 # ---- EXISTING ROUTES ----
